Praktikum Pertemuan 6

+++ app/Http/Controllers/StudentController.php
@@ -12,6 +12,15 @@ class StudentController extends Controller
     public function index()
     {
         $students = Student::all();
+
+        //jika data kosong maka kirim status code 204
+        if($students->isEmpty()){
+            $data = [
+                'message' => 'Resource Is Empty'
+            ];
+
+            return response()->json($data, 204);
+        }
         
         $data = [
             'message' => 'Get all students',
@@ -23,6 +32,15 @@ class StudentController extends Controller
     }
 
     public function store(Request $request){
+
+        //validasi data request
+        $request->validate([
+            'nama' => 'required',
+            'nim' => 'required',
+            'email' => 'required|email',
+            'jurusan' => 'required'
+        ]);
+
         $input = [
             'nama' => $request->nama,
             'nim' => $request->nim,
@@ -43,29 +61,63 @@ class StudentController extends Controller
     public function update(Request $request , $id){
         $student = Student::find($id);
 
-        $input = [
-            'nama' => $request->nama,
-            'nim' => $request->nim,
-            'email' => $request->email,
-            'jurusan' => $request->jurusan
-        ];
 
-        $student->update($input);
+        if(!$student){
+            $data = [
+                'message' => 'Student not found'
+            ];
+
+            return response()->json($data, 404);
+        }
+
+        $student->update([
+            'nama' => $request->nama ?? $student->nama,
+            'nim' => $request->nim ?? $student->nim,
+            'email' => $request->email ?? $student->email,
+            'jurusan' => $request->jurusan ?? $student->jurusan
+        ]);
 
         $data = [
-            'message' => 'Student is updated succesfully',
+            'message' => 'Data is updated succesfully',
             'data' => $student
         ];
-            return response()->json($data, 404);
+            return response()->json($data, 200);
     }
 
     public function destroy($id){
         $student = Student::find($id);
 
+        if(!$student){
+            $data = [
+                'message' => 'Student not found'
+            ];
+
+            return response()->json($data, 404);
+        }
+
         $student->delete();
 
         $data = [
             'message' => 'Student is deleted succesfully',
+            'data' => $student
+        ];
+
+        return response()->json($data, 200);
+    }
+
+    public function show($id){
+        $student = Student::find($id);
+
+        if(!$student){
+            $data = [
+                'message' => 'Student not found'
+            ];
+
+            return response()->json($data, 404);
+        }
+
+        $data = [
+            'message' => 'Show Detail Resource',
             'data' => $student
         ];
